Prepare for remediation commits
This commit moves some things around in preparation for adding support for remediation commits. It
does a number of things which will be used in subsequent commits:

* Adds config file support for remediation commits, which are disabled by default.
* Updates the DCO check function call to be aware of remediation commits.
* Refactors signoff failure feedback into a single function, which will be subsequently expanded.
* Update the snapshot to reflect the updated text.

In addition, this commit updates the default feedback text. While testing I noted that there could
be situations where amending a commit would not fix the issue (e.g., if the failing commit was not
at the tip). Since both amend and rebase regenerate the hash and require a force push, I
consolidated feedback for both into one common text.

Signed-off-by: Brian Warner <brian@bdwarner.com>